refactor(theme): move useTheme hook to its own file and fix type-only imports
