+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
-import { useTheme } from "@/context/ThemeContext";
+import { useTheme } from "@/hooks";
 import { Footer, Header } from "@/components";
 import TracksPage from "@/pages/TracksPage";
 import "react-toastify/dist/ReactToastify.css";
