+++ client/src/components/layout/Header/ThemeToggle.tsx
@@ -1,5 +1,5 @@
 import { Sun, Moon } from "lucide-react";
-import { useTheme } from "@/context/ThemeContext";
+import { useTheme } from "@/hooks";
 import { getValidClassNames } from "@/helpers";
 
 export function ThemeToggle() {
