+++ client/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export { useTheme } from "./useTheme";
 export { useTracks } from "./useTracks";
 export { useGenres } from "./useGenres";
 export { useArtists } from "./useArtists";
