+++ client/src/hooks/useTheme.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { ThemeContext } from "@/context/ThemeContext";
+
+export function useTheme() {
+  const ctx = useContext(ThemeContext);
+  if (!ctx) throw new Error("useTheme must be inside ThemeProvider");
+  return ctx;
+}
